rename .html files to .svelte
