+++ test/apps/basics/src/routes/[slug].svelte
