+++ test/apps/basics/src/routes/_error.svelte
