+++ test/apps/basics/src/routes/a.svelte
