+++ test/apps/basics/src/routes/b/index.svelte
