+++ test/apps/basics/src/routes/const.svelte
