+++ test/apps/basics/src/routes/delete-test/index.svelte
