+++ test/apps/basics/src/routes/dirs/bar/index.svelte
