+++ test/apps/basics/src/routes/dirs/foo/index.svelte
