+++ test/apps/basics/src/routes/echo-query/index.svelte
