+++ test/apps/basics/src/routes/index.svelte
