+++ test/apps/basics/src/routes/unsafe-replacement.svelte
