+++ test/apps/css/src/routes/foo.svelte
