+++ test/apps/errors/src/routes/_error.svelte
