+++ test/apps/errors/src/routes/index.svelte
