+++ test/apps/errors/src/routes/throw.svelte
