+++ test/apps/export/src/routes/index.svelte
