+++ test/apps/ignore/src/routes/[slug].svelte
