+++ test/apps/ignore/src/routes/index.svelte
