+++ test/apps/layout/src/routes/[slug].svelte
