+++ test/apps/preloading/src/routes/_layout.svelte
