+++ test/apps/preloading/src/routes/foo.svelte
