+++ test/apps/preloading/src/routes/index.svelte
