+++ test/apps/preloading/src/routes/slow-preload.svelte
