+++ test/apps/redirects/src/routes/_error.svelte
