+++ test/apps/redirects/src/routes/index.svelte
