+++ test/apps/redirects/src/routes/redirect-to-external.svelte
