+++ test/apps/redirects/src/routes/redirect-to-root.svelte
