+++ test/apps/with-basepath/src/routes/_error.svelte
