+++ test/apps/with-basepath/src/routes/index.svelte
