+++ test/apps/with-sourcemaps-webpack/src/routes/_error.svelte
