+++ test/apps/with-sourcemaps-webpack/src/routes/index.svelte
