+++ test/apps/with-sourcemaps/src/routes/_error.svelte
