+++ test/apps/with-sourcemaps/src/routes/index.svelte
